refactor(runtimed): extract 18 non-LaunchKernel request handlers
Moves every match arm from `handle_notebook_request` except LaunchKernel
out into `crates/runtimed/src/requests/*.rs`. Each handler is a
`pub(crate) async fn handle(...)` that takes exactly the state it needs
and returns a `NotebookResponse`. Lock scoping, log lines, and error
strings are preserved verbatim.

Helpers consumed across the dispatcher boundary have been lifted to
`pub(crate)`: `check_and_broadcast_sync_state`, `check_and_update_trust_state`,
`detect_room_runtime`, `format_source`, `formatter_actor`,
`clone_notebook_to_disk`, `canonical_target_path`, `try_claim_path`,
`save_notebook_to_disk`, `finalize_untitled_promotion`,
`format_notebook_cells`, `handle_sync_environment`, and the `SaveError` enum.

LaunchKernel (still ~1340 lines) remains inline — extracted in the next
commit so its dep surface can be reviewed separately.
