+++ crates/runtimed/src/requests/check_tool_available.rs
@@ -1,3 +1,11 @@
 //! `NotebookRequest::CheckToolAvailable` handler.
-//!
-//! Placeholder module; the handler is extracted in a follow-up commit.
+
+use crate::protocol::NotebookResponse;
+
+pub(crate) async fn handle(tool: String) -> NotebookResponse {
+    let available = match tool.as_str() {
+        "deno" => kernel_launch::tools::check_deno_available_without_bootstrap().await,
+        _ => false,
+    };
+    NotebookResponse::ToolAvailable { available }
+}
