+++ crates/runtimed/src/requests/get_doc_bytes.rs
@@ -1,3 +1,10 @@
 //! `NotebookRequest::GetDocBytes` handler.
-//!
-//! Placeholder module; the handler is extracted in a follow-up commit.
+
+use crate::notebook_sync_server::NotebookRoom;
+use crate::protocol::NotebookResponse;
+
+pub(crate) async fn handle(room: &NotebookRoom) -> NotebookResponse {
+    let mut doc = room.doc.write().await;
+    let bytes = doc.save();
+    NotebookResponse::DocBytes { bytes }
+}
