+++ crates/runtimed/src/requests/get_metadata_snapshot.rs
@@ -1,3 +1,12 @@
 //! `NotebookRequest::GetMetadataSnapshot` handler.
-//!
-//! Placeholder module; the handler is extracted in a follow-up commit.
+
+use crate::notebook_sync_server::NotebookRoom;
+use crate::protocol::NotebookResponse;
+
+pub(crate) async fn handle(room: &NotebookRoom) -> NotebookResponse {
+    let doc = room.doc.read().await;
+    let snapshot = doc
+        .get_metadata_snapshot()
+        .and_then(|s| serde_json::to_string(&s).ok());
+    NotebookResponse::MetadataSnapshot { snapshot }
+}
