+++ crates/runtimed/src/requests/get_raw_metadata.rs
@@ -1,3 +1,10 @@
 //! `NotebookRequest::GetRawMetadata` handler.
-//!
-//! Placeholder module; the handler is extracted in a follow-up commit.
+
+use crate::notebook_sync_server::NotebookRoom;
+use crate::protocol::NotebookResponse;
+
+pub(crate) async fn handle(room: &NotebookRoom, key: String) -> NotebookResponse {
+    let doc = room.doc.read().await;
+    let value = doc.get_metadata(&key);
+    NotebookResponse::RawMetadata { value }
+}
