+++ crates/runtimed/src/requests/shutdown_kernel.rs
@@ -1,3 +1,36 @@
 //! `NotebookRequest::ShutdownKernel` handler.
-//!
-//! Placeholder module; the handler is extracted in a follow-up commit.
+
+use crate::notebook_sync_server::{send_runtime_agent_request, NotebookRoom};
+use crate::protocol::NotebookResponse;
+
+pub(crate) async fn handle(room: &NotebookRoom) -> NotebookResponse {
+    // Send shutdown RPC but keep the runtime agent alive — it stays
+    // connected for potential RestartKernel. The kernel process dies
+    // but the runtime agent subprocess and socket connection remain.
+    let has_runtime_agent = room.runtime_agent_request_tx.lock().await.is_some();
+    if has_runtime_agent {
+        let _ = send_runtime_agent_request(
+            room,
+            notebook_protocol::protocol::RuntimeAgentRequest::ShutdownKernel,
+        )
+        .await;
+        // Keep runtime agent alive (runtime_agent_handle + runtime_agent_request_tx stay set)
+        // so LaunchKernel can send RestartKernel. ExecuteCell/RunAllCells
+        // check kernel.status from RuntimeStateDoc and return NoKernel
+        // when status is "shutdown".
+        //
+        // Update RuntimeStateDoc to reflect shutdown
+        {
+            let mut sd = room.state_doc.write().await;
+            let mut changed = false;
+            changed |= sd.set_kernel_status("shutdown");
+            changed |= sd.set_queue(None, &[]);
+            if changed {
+                let _ = room.state_changed_tx.send(());
+            }
+        }
+        NotebookResponse::KernelShuttingDown {}
+    } else {
+        NotebookResponse::NoKernel {}
+    }
+}
